pyproject.toml
[tool.poetry]
name = "nkoe"
version = "0.1.0"
description = "A basic aiohttp server."
authors = ["Serban Senciuc <serban.senciuc@gmail.com>"]
repository = "https://github.com/senciucserban/nkoe"
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.7"
aiohttp = "^3.5"
python-dotenv = "^0.10.3"
coloredlogs = "^10.0"
pytz = "^2019.2"
toml = "^0.10.0"
jara = "^1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
server = 'manager:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
